+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## To Be Released
 
 * feat(apps):  add the `private_network_ids` field to the model
-* feat(database/users): add the `DatabaseUserPasswordReset` method
+* feat(database/users): add the `DatabaseUserResetPassword` method
 
 ## 7.0.0